+++ doc/release_notes.md
@@ -5,6 +5,11 @@
 1.  All of the database type constants have been made references in order to
     avoid the necssity of taking a reference (or a double reference when
     binding the type directly).
+1.  Added methods [Metadata::data_type()](crate::Metadata::data_type()),
+    [Metadata::is_sparse_vector()](crate::Metadata::is_sparse_vector()),
+    [Metadata::vector_dimensions()](crate::Metadata::vector_dimensions()) and
+    [Metadata::vector_dimensions()](crate::Metadata::vector_storage_format())
+    and the enumeration [VectorStorageFormat](crate::VectorStorageFormat).
 1.  Eliminated hang when an error occurs during a DML returning statemnt
     ([issue 16](https://github.com/oracle/rust-oracledb/issues/16)).